requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
numpy==2.0.2
scipy==1.13.1
pydantic==2.8.2
cachetools==5.3.3
pytest==8.3.2

